# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: BedLink API contract
  version: 1.1.0
  description: Step 1 contract with Step 3 availability implementation; workflow operations remain planned. JSON uses camelCase. REST with Supabase Auth bearer
    access tokens. Server time and inventory are authoritative. Read decisions.md and frontend-handoff.md for normative
    cross-resource invariants, ranking, idempotency and lifecycle rules. No patient names or clinical notes. Unknown
    input fields/queries are 400. All bodies are application/json; every response includes serverTime. All timestamps
    end in Z. API deployment URL remains configurable.
servers:
- url: /api/v1
  description: Relative base path on the configured API origin; no live service implied.
security:
- bearerAuth: []
...
paths:
...
  /hospitals:
    get:
      operationId: listHospitals
      summary: List hospitals with all bed pools
      tags:
      - Hospitals
      description: 'Returns the complete demo catalog. No query parameters. Reading never verifies availability.
        Authorization: Nurse or dispatcher may read the configured hospital catalog. No patient records are included.
        All errors use ErrorResponse.'
      security:
      - bearerAuth: []
      x-roles:
      - dispatcher
      - nurse
      x-authorization: Nurse or dispatcher may read the configured hospital catalog. No patient records are included.
      parameters: []
      responses:
        '200':
          description: Success.
          headers:
            Cache-Control:
              description: Always no-store. Do not serve cached capacity or request state.
              schema:
                type: string
                const: no-store
            Date:
              description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalsResponse'
              examples:
                01-success-01-catalog:
                  externalValue: ./examples/01-success/01-catalog.response.json
                04-no-eligible-00-catalog:
                  externalValue: ./examples/04-no-eligible/00-catalog.response.json
                05-stale-00-catalog:
                  externalValue: ./examples/05-stale/00-catalog.response.json
                06-capacity-unavailable-00-catalog:
                  externalValue: ./examples/06-capacity-unavailable/00-catalog.response.json
                08-last-bed-00-catalog:
                  externalValue: ./examples/08-last-bed/00-catalog.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /hospitals/{hospitalId}/bed-pools:
    get:
      operationId: listBedPools
      summary: List pools for one hospital
      tags:
      - Hospitals
      description: 'Returns current pool versions, counts, and freshness. No query parameters. Authorization: Nurse
        or dispatcher may read the configured hospital catalog. No patient records are included. All errors use
        ErrorResponse.'
      security:
      - bearerAuth: []
      x-roles:
      - dispatcher
      - nurse
      x-authorization: Nurse or dispatcher may read the configured hospital catalog. No patient records are included.
      parameters:
      - $ref: '#/components/parameters/hospitalId'
      responses:
        '200':
          description: Success.
          headers:
            Cache-Control:
              description: Always no-store. Do not serve cached capacity or request state.
              schema:
                type: string
                const: no-store
            Date:
              description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BedPoolsResponse'
              examples:
                07-nurse-conflict-01-read:
                  externalValue: ./examples/07-nurse-conflict/01-read.response.json
                07-nurse-conflict-04-refetch:
                  externalValue: ./examples/07-nurse-conflict/04-refetch.response.json
                10-hold-expiry-03-inventory:
                  externalValue: ./examples/10-hold-expiry/03-inventory.response.json
                11-cancel-hold-02-inventory:
                  externalValue: ./examples/11-cancel-hold/02-inventory.response.json
                12-cancel-request-02-inventory:
                  externalValue: ./examples/12-cancel-request/02-inventory.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '404':
          $ref: '#/components/responses/NotFound'
  /hospitals/{hospitalId}/bed-pools/{bedPoolId}:
    patch:
      operationId: updateBedPool
      summary: Update or explicitly verify free beds
      tags:
      - Hospitals
      description: 'Compare supplied version and enforce activeHoldCount <= reportedFreeBeds <= capacity atomically.
        update and verify refresh verifiedAt and inventoryUpdatedAt to serverTime and increment version, including
        same-count updates. verify requires the supplied count to equal the current count. Reads and replay never
        verify. Authorization: Nurse only; server-managed hospital assignment must equal hospitalId and the resource
        must belong to that hospital. All errors use ErrorResponse.'
      security:
      - bearerAuth: []
      x-roles:
      - nurse
      x-authorization: Nurse only; server-managed hospital assignment must equal hospitalId and the resource must
        belong to that hospital.
      parameters:
      - $ref: '#/components/parameters/hospitalId'
      - $ref: '#/components/parameters/bedPoolId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Success.
          headers:
            Cache-Control:
              description: Always no-store. Do not serve cached capacity or request state.
              schema:
                type: string
                const: no-store
            Date:
              description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
              schema:
                type: string
            Idempotency-Replayed:
              description: true on replay; false on newly executed and stored result.
              schema:
                type: boolean
            Idempotency-Expires-At:
              description: End of the 24-hour replay guarantee. Present when an idempotency result has been stored.
              schema:
                $ref: '#/components/schemas/Timestamp'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BedPoolResponse'
              examples:
                07-nurse-conflict-02-first-update:
                  externalValue: ./examples/07-nurse-conflict/02-first-update.response.json
                07-nurse-conflict-05-explicit-verify:
                  externalValue: ./examples/07-nurse-conflict/05-explicit-verify.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: JSON body exceeds the configured request-size limit. Error code INVALID_INPUT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '409':
          $ref: '#/components/responses/Conflict'
      x-conflict-codes:
      - VERSION_CONFLICT
      - FREE_COUNT_BELOW_HOLDS
      - VERIFICATION_COUNT_CHANGED
      - IDEMPOTENCY_KEY_REUSED
      - IDEMPOTENCY_IN_PROGRESS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityUpdate'
            examples:
              07-nurse-conflict-02-first-update:
                externalValue: ./examples/07-nurse-conflict/02-first-update.request.json
              07-nurse-conflict-03-conflicting-update:
                externalValue: ./examples/07-nurse-conflict/03-conflicting-update.request.json
              07-nurse-conflict-05-explicit-verify:
                externalValue: ./examples/07-nurse-conflict/05-explicit-verify.request.json
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token. Backend validates signature, issuer, audience, expiry and current
        authorization. Roles/assignments from trusted server-managed records, never client-editable user metadata.
        Never use a Supabase service-role key in a frontend.
  parameters:
    hospitalId:
      name: hospitalId
      in: path
      required: true
      description: Opaque hospitalId. Must belong to the parent path resource.
      schema:
        $ref: '#/components/schemas/Id'
    bedPoolId:
      name: bedPoolId
      in: path
      required: true
      description: Opaque bedPoolId. Must belong to the parent path resource.
      schema:
        $ref: '#/components/schemas/Id'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: New UUID or similarly unique key for each user intent. Reuse unchanged on retries. Scoped to
        authenticated issuer+subject across all mutations; retained 24 hours. See frontend-handoff.md.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
  responses:
    BadRequest:
      description: INVALID_INPUT or IDEMPOTENCY_KEY_REQUIRED; malformed JSON, invalid parameters or unknown fields.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: UNAUTHENTICATED; absent, invalid, expired, or revoked bearer token.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer challenge.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: FORBIDDEN; authenticated role cannot use this operation.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: NOT_FOUND; resource absent, wrong parent, or object outside the caller authorization scope. Do
        not reveal which.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Business or concurrency conflict. Inspect error.code; operation-specific codes are listed on
        the operation.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
        Retry-After:
          description: Seconds before retry. Required for 429/503; present for IDEMPOTENCY_IN_PROGRESS (1 second).
          schema:
            type: integer
            minimum: 1
        Idempotency-Replayed:
          description: true on replay; false on newly executed and stored result.
          schema:
            type: boolean
        Idempotency-Expires-At:
          description: End of the 24-hour replay guarantee. Present when an idempotency result has been stored.
          schema:
            $ref: '#/components/schemas/Timestamp'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: 'UNSUPPORTED_MEDIA_TYPE; JSON body requires Content-Type: application/json.'
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: RATE_LIMITED; honor Retry-After seconds.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
        Retry-After:
          description: Seconds before retry. Required for 429/503; present for IDEMPOTENCY_IN_PROGRESS (1 second).
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: INTERNAL_ERROR; no internal stack or database details.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: SERVICE_UNAVAILABLE; result may be unknown, retry a mutation with the same Idempotency-Key.
      headers:
        Cache-Control:
          description: Always no-store. Do not serve cached capacity or request state.
          schema:
            type: string
            const: no-store
        Date:
          description: Current HTTP response generation time; use with immediate refetch after idempotent replay.
          schema:
            type: string
        Retry-After:
          description: Seconds before retry. Required for 429/503; present for IDEMPOTENCY_IN_PROGRESS (1 second).
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Id:
      type: string
      minLength: 1
      maxLength: 128
      description: Opaque identifier. Never parse it; URL-encode when used in paths.
    Timestamp:
      type: string
      format: date-time
      pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,3})?Z$
      description: UTC ISO 8601 timestamp, with Z suffix. Assigned by the server.
    Location:
      type: object
      additionalProperties: false
      required:
      - latitude
      - longitude
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
    Resource:
      type: string
      enum:
      - icu
      - ventilator
      - oxygen
    Specialty:
      type: string
      enum:
      - cardiac
      - burns
...
    HospitalSummary:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - location
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          minLength: 1
          maxLength: 120
        location:
          $ref: '#/components/schemas/Location'
    FreshnessPolicy:
      type: object
      additionalProperties: false
      required:
      - policyVersion
      - agingAfterMinutes
      - staleAfterMinutes
      properties:
        policyVersion:
          type: string
          minLength: 1
        agingAfterMinutes:
          type: number
          minimum: 0
        staleAfterMinutes:
          type: number
          exclusiveMinimum: 0
      description: Effective server thresholds; agingAfterMinutes < staleAfterMinutes. Returned on pools so catalog
        and nurse screens do not hardcode demo defaults.
    BedPool:
      type: object
      additionalProperties: false
      required:
      - id
      - hospitalId
      - label
      - resources
      - specialties
      - capacity
      - reportedFreeBeds
      - activeHoldCount
      - availableBeds
      - loadRatio
      - verifiedAt
      - dataAgeMinutes
      - freshness
      - freshnessPolicy
      - inventoryUpdatedAt
      - version
      properties:
        id:
          $ref: '#/components/schemas/Id'
        hospitalId:
          $ref: '#/components/schemas/Id'
        label:
          type: string
          minLength: 1
          maxLength: 120
        resources:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
          minItems: 1
          maxItems: 3
          uniqueItems: true
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/Specialty'
          maxItems: 2
          uniqueItems: true
        capacity:
          type: integer
          minimum: 1
        reportedFreeBeds:
          type: integer
          minimum: 0
          description: Physical unoccupied staffed usable beds, INCLUDING beds held by BedLink. Never subtract holds
            in this nurse-entered value.
        activeHoldCount:
          type: integer
          minimum: 0
          description: Unexpired active BedLink holds. Pending attempts consume no capacity.
        availableBeds:
          type: integer
          minimum: 0
          description: reportedFreeBeds - activeHoldCount. Eligible for new holds only if fresh enough.
        loadRatio:
          type: number
          minimum: 0
          maximum: 1
          description: 1 - availableBeds / capacity; pool occupancy plus BedLink reservations, NOT ED queue length
            or hospital-wide utilization.
        verifiedAt:
          anyOf:
          - $ref: '#/components/schemas/Timestamp'
          - type: 'null'
        dataAgeMinutes:
          anyOf:
          - type: number
            minimum: 0
            description: (serverTime - verifiedAt) / 60000, fractional minutes. Null when never verified.
          - type: 'null'
        freshness:
          type: string
          enum:
          - fresh
          - aging
          - stale
          - unverified
        freshnessPolicy:
          $ref: '#/components/schemas/FreshnessPolicy'
        inventoryUpdatedAt:
          $ref: '#/components/schemas/Timestamp'
        version:
          type: integer
          minimum: 1
          description: Monotonic version incremented on verification, count edits, and hold/inventory changes. Reads
            never increment it.
      description: Each physical bed belongs to exactly one pool. 0 <= activeHoldCount <= reportedFreeBeds <= capacity.
        verifiedAt changes only on successful nurse update or explicit verification; arrival and hold changes do
        not refresh it.
      allOf:
      - if:
          properties:
            freshness:
              const: unverified
        then:
          properties:
            verifiedAt:
              type: 'null'
            dataAgeMinutes:
              type: 'null'
        else:
          properties:
            verifiedAt:
              $ref: '#/components/schemas/Timestamp'
            dataAgeMinutes:
              type: number
              minimum: 0
    Hospital:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - location
      - bedPools
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          minLength: 1
          maxLength: 120
        location:
          $ref: '#/components/schemas/Location'
        bedPools:
          type: array
          items:
            $ref: '#/components/schemas/BedPool'
    HospitalsResponse:
      type: object
      additionalProperties: false
      required:
      - serverTime
      - hospitals
      properties:
        serverTime:
          $ref: '#/components/schemas/Timestamp'
        hospitals:
          type: array
          items:
            $ref: '#/components/schemas/Hospital'
      description: Entire configured demo catalog, ordered by hospital ID and pool ID using ascending Unicode code-point
        order. No pagination in Step 1.
    BedPoolsResponse:
      type: object
      additionalProperties: false
      required:
      - serverTime
      - hospital
      - bedPools
      properties:
        serverTime:
          $ref: '#/components/schemas/Timestamp'
        hospital:
          $ref: '#/components/schemas/HospitalSummary'
        bedPools:
          type: array
          items:
            $ref: '#/components/schemas/BedPool'
    BedPoolResponse:
      type: object
      additionalProperties: false
      required:
      - serverTime
      - bedPool
      properties:
        serverTime:
          $ref: '#/components/schemas/Timestamp'
        bedPool:
          $ref: '#/components/schemas/BedPool'
    AvailabilityUpdate:
      type: object
      additionalProperties: false
      required:
      - operation
      - reportedFreeBeds
      - version
      properties:
        operation:
          type: string
          enum:
          - update
          - verify
        reportedFreeBeds:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
      description: All fields required. update records a new count; verify explicitly confirms the unchanged current
        count. verify with a different count fails 409 VERIFICATION_COUNT_CHANGED. Both refresh verifiedAt and increment
        version once; replay does neither. Count must be between activeHoldCount and capacity.
...
    FieldDetail:
      type: object
      additionalProperties: false
      required:
      - field
      - code
      - message
      properties:
        field:
          type: string
          minLength: 1
          description: JSON Pointer for a body field, or header name.
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    Error:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_INPUT
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - UNSUPPORTED_MEDIA_TYPE
          - RATE_LIMITED
          - INTERNAL_ERROR
          - SERVICE_UNAVAILABLE
          - IDEMPOTENCY_KEY_REQUIRED
          - IDEMPOTENCY_KEY_REUSED
          - IDEMPOTENCY_IN_PROGRESS
          - VERSION_CONFLICT
          - FREE_COUNT_BELOW_HOLDS
          - VERIFICATION_COUNT_CHANGED
          - REQUEST_NOT_SEARCHING
          - HOSPITAL_ALREADY_ATTEMPTED
          - POOL_REQUIREMENTS_NOT_MET
          - CAPACITY_UNAVAILABLE
          - AVAILABILITY_OUTDATED
          - ATTEMPT_DEADLINE_PASSED
          - ATTEMPT_NOT_PENDING
          - HOLD_EXPIRED
          - HOLD_NOT_ACTIVE
          - REQUEST_TERMINAL
        message:
          type: string
          minLength: 1
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldDetail'
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - serverTime
      - error
      properties:
        serverTime:
          $ref: '#/components/schemas/Timestamp'
        error:
          $ref: '#/components/schemas/Error'